+++ src/pages/Signup/SignupStep2.tsx
@@ -1,10 +1,9 @@
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import InputField from "./InputField";
 import { useLocation } from "react-router-dom";
 import { MoreTerms } from "../../assets/svg";
-import { verifyCode, verifyPhoneNumber } from "../../apis/verify/verify";
-import { AxiosError } from "axios";
 import { useSignUpMutation } from "../../hooks/useSignup";
+import PhoneVerification from "../../components/PhoneVerification";
 
 interface Agreements {
   all: boolean;
@@ -20,12 +19,7 @@ function SignupStep2() {
 
   const [name, setName] = useState("");
   const [phoneNumber, setPhoneNumber] = useState("");
-  const [certificationCode, setCertificationCode] = useState("");
-  const [isCodeSent, setIsCodeSent] = useState(false);
   const [isCodeConfirmed, setIsCodeConfirmed] = useState(false);
-  const [timer, setTimer] = useState(180); // 3분 타이머
-  const [isTimerRunning, setIsTimerRunning] = useState(false);
-  const [codeError, setCodeError] = useState("");
 
   const [agreements, setAgreements] = useState<Agreements>({
     all: false,
@@ -54,62 +48,6 @@ function SignupStep2() {
       updated.all = Object.values(updated).every(Boolean);
       return updated;
     });
-  };
-
-  /** 타이머 시작 */
-  useEffect(() => {
-    let timerInterval: ReturnType<typeof setInterval> | null = null;
-
-    if (isTimerRunning && timer > 0) {
-      timerInterval = setInterval(() => {
-        setTimer((prev) => prev - 1);
-      }, 1000);
-    } else if (timer === 0) {
-      if (timerInterval) {
-        clearInterval(timerInterval);
-      }
-      setIsTimerRunning(false);
-    }
-
-    return () => {
-      if (timerInterval) {
-        clearInterval(timerInterval);
-      }
-    };
-  }, [isTimerRunning, timer]);
-
-  /** 휴대폰 번호 확인 */
-  const validatePhoneNumber = () => {
-    const regex = /^01[0-9]{8,9}$/; // 01로 시작하는 10-11자리 숫자
-    return regex.test(phoneNumber);
-  };
-
-  /** 인증하기 버튼 핸들러 */
-  const handleSendCode = async () => {
-    if (validatePhoneNumber()) {
-      await verifyCode(phoneNumber);
-      setIsCodeSent(true);
-      setIsTimerRunning(true);
-      setTimer(180); // 3분 타이머 시작
-    }
-  };
-
-  /** 인증번호 확인 */
-  const handleVerifyCode = async() => {
-    if (certificationCode.length === 6) {
-      try {
-        await verifyPhoneNumber({phoneNumber, certificationCode});
-        setCodeError("");
-        setIsCodeConfirmed(true);
-        setIsTimerRunning(false);
-      } catch (error) {
-        if (error instanceof AxiosError) {
-          setCodeError(error.response?.data?.message || "인증에 실패했습니다.");
-        } else {
-          setCodeError("인증에 실패했습니다.");
-        }
-      }
-    }
   };
 
   const isFormValid =
@@ -149,72 +87,8 @@ function SignupStep2() {
           />
         </div>
 
-          {/* 휴대폰 입력창 */}
-          <div className="w-full flex flex-col gap-[10px]">
-            <label htmlFor="phone" className="text-[16px] font-medium text-black-700">
-              휴대폰 번호
-            </label>
-            <div className="flex items-center gap-[12px]">
-              {/* 휴대폰 입력창 */}
-              <div className="flex items-center h-[50px] px-[20px] border border-gray-400 rounded-[5px] flex-1 gap-[20px] relative">
-                <input
-                  type="text"
-                  placeholder="휴대폰 번호를 -없이 입력해주세요"
-                  value={phoneNumber}
-                  onChange={(e) => setPhoneNumber(e.target.value)}
-                  className="w-full outline-none text-[14px] font-medium placeholder-gray-400"
-                />
-              </div>
-              <button
-                onClick={handleSendCode}
-                disabled={!validatePhoneNumber()}
-                className={`h-[50px] px-[20px] rounded-[5px] text-[15px] font-medium ${
-                  validatePhoneNumber()
-                    ? "bg-blue-500 text-white-100 hover:bg-blue-400"
-                    : "bg-blue-250 text-white-100"
-                }`}
-              >
-                인증하기
-              </button>
-            </div>
-
-            {/* 인증번호 입력창 */}
-            {isCodeSent && !isCodeConfirmed && (
-              <div className="flex items-center gap-[12px] mt-[10px]">
-                <div className="flex items-center h-[50px] px-[20px] border border-gray-400 rounded-[5px] flex-1 relative">
-                  <input
-                    type="text"
-                    placeholder="인증번호를 입력해주세요"
-                    value={certificationCode}
-                    onChange={(e) => setCertificationCode(e.target.value)}
-                    className="w-full outline-none text-[14px] font-medium placeholder-gray-400"
-                  />
-                  <span className="text-red-500 text-[14px] absolute right-[15px]">
-                    {`${Math.floor(timer / 60)}:${String(timer % 60).padStart(2, "0")}`}
-                  </span>
-                </div>
-                <button
-                  onClick={handleVerifyCode}
-                  disabled={certificationCode.length !== 6}
-                  className={`h-[50px] px-[20px] rounded-[5px] text-[15px] font-medium ${
-                    certificationCode.length === 6
-                      ? "bg-blue-500 text-white-100 hover:bg-blue-400"
-                      : "bg-blue-250 text-white-100"
-                  }`}
-                >
-                  확인하기
-                </button>
-              </div>
-            )}
-
-            {/* 인증번호 오류 메시지 */}
-            {codeError && <span className="text-red-500 text-[13px] mt-[10px]">{codeError}</span>}
-            {/* 인증 완료 메시지 */}
-            {isCodeConfirmed && (
-              <span className="text-[13px] text-green-500 mt-[10px]">확인되었습니다.</span>
-            )}
-          </div>
-        </div>
+        <PhoneVerification phoneNumber={phoneNumber} setPhoneNumber={setPhoneNumber} onVerifySuccess={() => setIsCodeConfirmed(true)} />
+      </div>
 
         {/* 약관 동의 섹션 */}
         <div className="w-full mb-[27px]">
